DAOController: c6. Add batching request to getActiveProposals & getInactiveProposals
